requirements.txt
argon2-cffi==21.1.0
asgiref==3.4.1
astroid==2.8.4
attrs==21.2.0
backports.entry-points-selectable==1.1.0
bandit==1.7.0
beautifulsoup4==4.10.0
black==21.9b0
boto3==1.19.3
botocore==1.22.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
crc32c==2.2.post0
cryptography==3.4.8
Deprecated==1.2.13
distlib==0.3.2
Django==3.2.7
django-cors-headers==3.8.0
django-filter==2.4.0
django-oauth-toolkit==1.5.0
django-storages==1.12.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
docopt==0.6.2
drf-spectacular==0.18.2
Faker==8.13.2
filelock==3.0.12
flake8==3.9.2
gitdb==4.0.7
GitPython==3.1.24
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
isort==5.9.3
jmespath==0.10.0
jsonschema==3.2.0
jwcrypto==1.0
kafka-python==2.0.2
lazy-object-proxy==1.6.0
lxml==4.6.3
lz4==3.1.3
mccabe==0.6.1
msgpack==1.0.2
mypy-extensions==0.4.3
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.4
pathspec==0.9.0
pbr==5.6.0
pip-api==0.0.22
pipreqs==0.4.10
platformdirs==2.3.0
pluggy==1.0.0
psycopg2==2.9.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pylint==2.11.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
smmap==4.0.0
soupsieve==2.2.1
sqlparse==0.4.2
stevedore==3.4.0
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
tox==3.24.4
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.8.0
wrapt==1.12.1
yarg==0.1.9
